+++ tests/testthat/test-as_duckplyr_tibble.R
@@ -1,0 +1,7 @@
+test_that("as_duckplyr_tibble() works", {
+  expect_s3_class(as_duckplyr_tibble(tibble(a = 1)), "duckplyr_df")
+  expect_equal(class(as_duckplyr_tibble(tibble(a = 1))), c("duckplyr_df", class(tibble())))
+
+  expect_s3_class(as_duckplyr_tibble(data.frame(a = 1)), "duckplyr_df")
+  expect_equal(class(as_duckplyr_tibble(data.frame(a = 1))), c("duckplyr_df", class(tibble())))
+})